manual correction in the documentation
The previous fold to 80 characters/line (max) introduced errors in
snippets of shell scripts; these are addressed.  In addition, lists
are reformatted, and direct links set as-such.

By now, only the principal head line causes a (the only) warning by
markdownlint (MD013, length of the line).  It is not edited because
it equally is the title of the publication about the tool.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>